refactor: use canonical algorithm imports
