+++ anomavision/algorithm/patchcore/__init__.py
@@ -1,5 +1,10 @@
 """Lightweight PatchCore anomaly detection algorithms."""
 
+from ..common.feature_extraction import ResnetEmbeddingsExtractor
 from .patchcore import PatchCore, build_patchcore_from_stats
 
-__all__ = ["PatchCore", "build_patchcore_from_stats"]
+__all__ = [
+    "PatchCore",
+    "ResnetEmbeddingsExtractor",
+    "build_patchcore_from_stats",
+]
