+++ tests/test_mahalanobis.py
@@ -1,7 +1,7 @@
 import pytest
 import torch
 
-from anomavision.mahalanobis import MahalanobisDistance
+from anomavision.algorithm.common.mahalanobis import MahalanobisDistance
 
 
 def make_identity_cov_inv(n: int, d: int) -> torch.Tensor:
