+++ tests/test_padim_lite.py
@@ -4,10 +4,10 @@ import torch
 
 def test_build_padim_from_stats_and_predict_cpu(make_stats, patch_extractor):
     stats, (N, D, W, H) = make_stats()
-    # patch extractor inside anomavision.padim_lite to our dummy
+    # patch extractor inside the canonical PaDiM-lite implementation
     patch_extractor(N, D, W, H)
 
-    from anomavision.padim_lite import build_padim_from_stats
+    from anomavision.algorithm.padim.padim_lite import build_padim_from_stats
 
     # build runtime model on CPU
     model = build_padim_from_stats(stats, device="cpu")
